Moved PasswordDialogue logic to a separate component
