+++ frontend/src/components/ChatRoomsList.vue
@@ -1,42 +1,10 @@
 <template>
   <div id="chatrooms-list">
-    <Dialog
-      header="Password required"
-      v-model:visible="displayPasswordDialog"
-      closeOnEscape
-      :style="{ width: '50vw' }"
-      :closable="false"
-    >
-      <div class="field p-fluid">
-        <small v-if="matchingPasswordError" id="password-help" class="p-error"
-          >Entered password is incorrect.</small
-        >
-        <small v-else id="password-help"
-          >Enter room password to join '{{ selectedRoomName }}'</small
-        >
-        <Password
-          id="password"
-          v-model="passwordValue"
-          :feedback="false"
-          showIcon="pi pi-eye"
-          toggleMask
-        />
-      </div>
-      <template #footer>
-        <Button
-          label="Cancel"
-          icon="pi pi-times"
-          @click="closePasswordDialog"
-          class="p-button-text"
-        />
-        <Button
-          label="OK"
-          icon="pi pi-check"
-          @click="validatePassword"
-          autofocus
-        />
-      </template>
-    </Dialog>
+    <ChatRoomPasswordDialogue
+      :isDialogVisible="displayPasswordDialog"
+      :roomName="selectedRoomName"
+      @update:isDialogVisible="displayPasswordDialog = $event"
+    />
     <DataTable
       :value="rooms"
       class="p-datatable-sm"
@@ -77,16 +45,14 @@
 </template>
 
 <script setup lang="ts">
-import { ref, inject, onMounted, defineEmits } from "vue";
+import { ref, inject, onMounted } from "vue";
 import { useRouter, useRoute } from "vue-router";
 import { Socket } from "socket.io-client";
 import RoomVisibility from "@/types/RoomVisibility";
+import ChatRoomPasswordDialogue from "./ChatRoomPasswordDialogue.vue";
 
 import DataTable from "primevue/datatable";
 import Column from "primevue/column";
-import Dialog from "primevue/dialog";
-import Password from "primevue/password";
-import Button from "primevue/button";
 
 const socket: Socket = inject("socketioInstance");
 
@@ -102,44 +68,22 @@ onMounted(() => {
 
 const router = useRouter();
 const route = useRoute();
+
 const displayPasswordDialog = ref(false);
-const selectedRoomName = ref(null);
-const selectedRoomIcon = ref(null);
+
+const selectedRoomName = ref("");
+// const selectedRoomIcon = ref("pi pi-hashtag");
+
 const onRowSelect = (event) => {
   if (event.data.password !== null) {
     displayPasswordDialog.value = true;
     selectedRoomName.value = event.data.name;
-    selectedRoomIcon.value = event.data.password
-      ? event.data.password
-      : event.data.visibility;
+    // selectedRoomIcon.value = "pi pi-hashtag"; // TODO: update
   } else {
-    router.push({ name: "ChatBox", params: { roomName: event.data.name } });
+    router.push({
+      name: "ChatBox",
+      params: { roomName: event.data.name },
+    });
   }
-};
-
-const passwordValue = ref();
-const closePasswordDialog = () => {
-  displayPasswordDialog.value = false;
-  passwordValue.value = null;
-  matchingPasswordError.value = false;
-};
-
-const matchingPasswordError = ref(false);
-const validatePassword = () => {
-  socket.emit("checkRoomPasswordMatch", {
-    name: selectedRoomName.value,
-    password: passwordValue.value,
-  });
-  socket.on("isRoomPasswordMatched", (isMatched) => {
-    if (isMatched) {
-      closePasswordDialog();
-      router.push({
-        name: "ChatBox",
-        params: { roomName: selectedRoomName.value },
-      });
-    } else {
-      matchingPasswordError.value = true;
-    }
-  });
 };
 </script>
